change `button.hpp` to `structs.hpp`.
