+++ 3ds/include/screens/musicMainScreen.hpp
@@ -27,7 +27,7 @@
 #include "screen.hpp"
 #include <string>
 #include <vector>
-#include "button.hpp"
+#include "structs.hpp"
 
 // Music Main.
 class MusicMain : public SCREEN
